Move auth implementation into a new 'auth' package
